% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addSVMPackage.R
\name{addSVMPackage}
\alias{addSVMPackage}
\title{add package to the bridge}
\usage{
addSVMPackage(method = NA, filePath = NA, softwarePath = NA,
  trainBinaryPath = NA, testBinaryPath = NA, wrapperPath = NA,
  verbose = FALSE)
}
\arguments{
\item{method}{name of solver}

\item{filePath}{path to wrapper}

\item{softwarePath}{path where to find the solver.}

\item{trainBinaryPath}{if not NULL this will override any search option}

\item{testBinaryPath}{if not NULL this will override any search option}

\item{wrapperPath}{if not NULL this will override any search option}

\item{verbose}{be verbose?}
}
\description{
add package to the bridge
}
\note{
first the given train and testBinaryPaths will be directly checked.
if the binary does not exist there, the softwarePath will be added and rechecked
and only if this does not work, the software will be searched via softwarePath.
so one can override the search by specifiying train-/testBinaryPath.
}

