% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSparseDataToConnection.R
\name{appendSparseDataToFile}
\alias{appendSparseDataToFile}
\title{appendSparseDataToFile}
\usage{
appendSparseDataToFile(filename, X, Y, verbose = FALSE, zeroBased = FALSE)
}
\arguments{
\item{filename}{file to write to (should NOT be opened by R)}

\item{X}{input data}

\item{Y}{labels}

\item{verbose}{be verbose?}

\item{zeroBased}{do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}
}
\description{
appendSparseDataToFile
}

