% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bashEscape.R
\name{bashEscape}
\alias{bashEscape}
\title{Escape a filename so it can be used on the command line.}
\usage{
bashEscape(s)
}
\arguments{
\item{s}{String/filename to escape.}
}
\value{
Escaped string.
}
\description{
Escape a filename so it can be used on the command line.
}
\note{
For now, only double quotes are added.
}

