% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkExecutionStrings.R
\name{checkExecutionStrings}
\alias{checkExecutionStrings}
\title{Check whether a binary's output contains given strings.}
\usage{
checkExecutionStrings(trainBinaryPath = NULL, patterns = NULL,
  applyKeyFix = FALSE, verbose = TRUE)
}
\arguments{
\item{trainBinaryPath}{Path of binary to check.}

\item{patterns}{List of patterns to check for. Checking is done via grep.}

\item{applyKeyFix}{Send a dummy key ('1') to the binary? This is necessary if the binary waits for a key prompt.}

\item{verbose}{Be verbose?}
}
\value{
True, if all patterns matches, False if not.
}
\description{
Given a pattern, this function will check whether the given binary has these patterns
in its output string, when called without any arguments. This helps all wrappers in identifying 
the correct binary.
}
\note{
To make sure that the binary is correct, it will be executed!
binary is really not exchangeable).
}

