% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkModel.R
\name{checkModel}
\alias{checkModel}
\title{Check model.}
\usage{
checkModel(model = NA, modelFile = NA, verbose = FALSE)
}
\arguments{
\item{model}{Model (in memory) to check. This cannot be used simultanously with modelFile.}

\item{modelFile}{Path to model (on disk) to check. . This cannot be used simultanously with model.}

\item{verbose}{Be verbose?}
}
\description{
Given a model, check it by calling the asking the corresponding wrapper.
}
\note{
As this is a basic for all other model readers, we export it.
}

