% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeOptimizationValues.R
\name{computeOptimizationValues}
\alias{computeOptimizationValues}
\title{Compute optimization values on given data for a model in LIBSVM 'format'.}
\usage{
computeOptimizationValues(X, Y, model, C = 0, verbose = FALSE)
}
\arguments{
\item{X}{Data}

\item{Y}{Labels}

\item{model}{LIBSVM model to work with.}

\item{C}{Regularization parameter.}

\item{verbose}{Report certain messages?}
}
\value{
A list of values with primal, dual, weight (0.5 w^T w) and training error.
}
\description{
Compute different optimization values like primal and dual for a given model.
}

