% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMWrapper.R
\name{createSVMWrapper}
\alias{createSVMWrapper}
\title{General class for any SVM wrapper.}
\usage{
createSVMWrapper(method, ...)
}
\arguments{
\item{method}{name of the class/object to create}
}
\description{
This will create an object of class SVMWrapper for a given name.
}
\note{
This function will first remove all member functions to make sure
	that no trash is lying around.
...
}

