% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMWrapper.R
\name{createSVMWrapperInternal}
\alias{createSVMWrapperInternal}
\title{General class for any SVM wrapper.}
\usage{
createSVMWrapperInternal(name, par.set, properties = character(0L),
  note = "")
}
\arguments{
\item{name}{Name of the class/object to create}

\item{par.set}{Parameter set (as a ParamHelpers class) for object}

\item{properties}{Properties in form of strings, e.g. multiclass, twoclass. See docs for more info.}

\item{note}{Notes for object, e.g. full name.}
}
\description{
This will create an object of class SVMWrapper for a given name.
}
\note{
This function will first remove all member functions to make sure
	that no trash is lying around.
}

