% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{createTestArguments.BSGD}
\alias{createTestArguments.BSGD}
\title{createTestArguments.BSGD}
\usage{
\method{createTestArguments}{BSGD}(x, testDataFile = "", modelFile = "",
  predictionOutput = "/dev/null", ...)
}
\arguments{
\item{x}{SVM object}

\item{testDataFile}{file to read test data from.}

\item{modelFile}{path to model, defaults to a temporary file (given by R).}

\item{predictionOutput}{path to where to put prediction output}
}
\value{
args		arguments for testing
}
\description{
createTestArguments.BSGD
}

