% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{createTestArguments.LIBSVM}
\alias{createTestArguments.LIBSVM}
\title{createTestArguments.LIBSVM}
\usage{
\method{createTestArguments}{LIBSVM}(x, testDataFile = "", modelFile = "",
  predictionsFile = "", ...)
}
\arguments{
\item{x}{SVM object}

\item{testDataFile}{file to read training data from.}

\item{modelFile}{path to model, defaults to a temporary file (given by R).}

\item{predictionsFile}{}
}
\description{
createTestArguments.LIBSVM
}

