% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{createTrainingArguments.BSGD}
\alias{createTrainingArguments.BSGD}
\title{createTrainingArguments.BSGD}
\usage{
\method{createTrainingArguments}{BSGD}(x, trainDataFile = "",
  modelFile = "", extraParameter = "", primalTime = 10, wallTime = 8 *
  60, cost = 1, gamma = 1, budget = 128, epochs = 1, ...)
}
\arguments{
\item{x}{SVM object}

\item{trainDataFile}{file to read training data from.}

\item{modelFile}{path to model, defaults to a temporary file (given by R).}

\item{extraParameter}{extra parameters for solver}

\item{primalTime}{}

\item{wallTime}{}

\item{cost}{cost parameter C.}

\item{gamma}{gamma parameter, note: RBF kernel used by pegasos is exp(-0.5 ...).}

\item{budget}{budget parameter.}

\item{epochs}{number of epochs to run.}
}
\value{
args		arguments for training
}
\description{
createTrainingArguments.BSGD
}

