% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{createTrainingArguments.LIBSVM}
\alias{createTrainingArguments.LIBSVM}
\title{Create training arguments for LIBSVM}
\usage{
\method{createTrainingArguments}{LIBSVM}(x, trainDataFile = "",
  modelFile = "", extraParameter = "", kernelCacheSize = 1024, cost = 1,
  svmType = "-1", useBias = FALSE, gamma = 1, epsilon = 0.001,
  degree = -1, coef0 = -1, nu = -1, shrinking = -1,
  probabilityEstimates = -1, weight = -1, n = -1, kernelType = "rbf",
  quietMode = FALSE, ...)
}
\arguments{
\item{x}{SVM object}

\item{trainDataFile}{file to read training data from.}

\item{modelFile}{path to model, defaults to a temporary file (given by R).}

\item{extraParameter}{extra parameters for solver}

\item{kernelCacheSize}{kernel cache parameter}

\item{cost}{cost parameter C.}

\item{svmType}{type of svm}

\item{useBias}{}

\item{gamma}{gamma parameter, note: RBF kernel used by pegasos is exp(-0.5 ...).}

\item{epsilon}{}

\item{degree}{}

\item{coef0}{}

\item{nu}{}

\item{shrinking}{}

\item{probabilityEstimates}{}

\item{weight}{}

\item{n}{}

\item{kernelType}{}

\item{quietMode}{}
}
\value{
args			arguments for training
}
\description{
Create training arguments for LIBSVM
}

