% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SVMperf_wrapper.R
\name{createTrainingArguments.SVMperf}
\alias{createTrainingArguments.SVMperf}
\title{evalSVMperf}
\usage{
\method{createTrainingArguments}{SVMperf}(x, trainDataFile = "",
  modelFile = "", kernelCacheSize = 1024, cost = 1, gamma = 1,
  k = 1000, epsilon = 0.001, saveFactor = -1, saveExponential = -1,
  modelPath = -1, wallTime = -1, primalTime = -1, extraParameter = "",
  ...)
}
\arguments{
\item{...}{parameters that will be passed on training and/or test callbacks
SVMperfTrainingParameterCallBack}

\item{...}{parameters that will be passed on training and/or test callbacks}
}
\description{
evalSVMperf
}

