% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{detectModel.BSGD}
\alias{detectModel.BSGD}
\title{Detect whether a file is a model for BSGD}
\usage{
\method{detectModel}{BSGD}(x, modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{x}{Object}

\item{modelFile}{File to check}

\item{verbose}{Verbose output?}
}
\value{
TRUE if the given modelFile exists and fits the BSGD model, or FALSE if not.
}
\description{
Detect whether a file is a model for BSGD
}
\note{
This is a very basic check, enough to distinguish the wrappers provided within the SVMBridge
}

