% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVM_wrapper.R
\name{detectModel.CVM}
\alias{detectModel.CVM}
\title{Detect whether a file is a model for CVM.}
\usage{
\method{detectModel}{CVM}(x, modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{x}{Object}

\item{modelFile}{File to check}

\item{verbose}{Verbose output?}
}
\value{
TRUE if the given modelFile exists and fits the CVM model, or FALSE if not.
}
\description{
Detect whether a file is a model for CVM.
}
\note{
This is a very basic check, enough to distinguish the wrappers provided within the SVMBridge
}

