% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASVM_wrapper.R
\name{detectModel.LASVM}
\alias{detectModel.LASVM}
\title{Detect whether a file is a model for LASVM.}
\usage{
\method{detectModel}{LASVM}(x, modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{x}{Object}

\item{modelFile}{File to check}

\item{verbose}{Verbose output?}
}
\value{
TRUE if the given modelFile exists and fits the LASVM model, or FALSE if not.
}
\description{
Detect whether a file is a model for LASVM.
}
\note{
This is a very basic check, enough to distinguish the wrappers provided within the SVMBridge
}

