% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{detectModel.LIBSVM}
\alias{detectModel.LIBSVM}
\title{Detect whether a file is a model for LIBSVM.}
\usage{
\method{detectModel}{LIBSVM}(x, modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{x}{Object}

\item{modelFile}{File to check}

\item{verbose}{Verbose output?}
}
\value{
TRUE if the given modelFile exists and fits the LIBSVM model, or FALSE if not.
}
\description{
Detect whether a file is a model for LIBSVM.
}
\note{
This is a very basic check, enough to distinguish the wrappers provided within the SVMBridge
}

