% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLSVM_wrapper.R
\name{detectModel.LLSVM}
\alias{detectModel.LLSVM}
\title{Detect whether a file is a model for LLSVM.}
\usage{
\method{detectModel}{LLSVM}(x, modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{x}{Object}

\item{modelFile}{File to check}

\item{verbose}{Verbose output?}
}
\value{
TRUE if the given modelFile exists and fits the LLSVM model, or FALSE if not.
}
\description{
Detect whether a file is a model for LLSVM.
}
\note{
This is a very basic check, enough to distinguish the wrappers provided within the SVMBridge
}

