% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparseFormat.R
\name{dumpSparseFormat}
\alias{dumpSparseFormat}
\title{dumpSparseFormat}
\usage{
dumpSparseFormat(labels, data)
}
\arguments{
\item{labels}{container for labels}

\item{data}{container for data}
}
\value{
sparseLine		string with all collected data
}
\description{
dumpSparseFormat
}

