% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseFormat.R
\name{dumpSparseLine}
\alias{dumpSparseLine}
\title{dumpSparseLine}
\usage{
dumpSparseLine(row)
}
\arguments{
\item{row}{array to read data from}
}
\value{
sparseLine		string with row data
}
\description{
dumpSparseLine
}

