% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandTilde.R
\name{expandTilde}
\alias{expandTilde}
\title{Expand path with tilde}
\usage{
expandTilde(path = NULL, verbose = FALSE)
}
\arguments{
\item{path}{path to expand}

\item{verbose}{print messages while searching?}
}
\description{
Given a path, try to expand the path in a crossplatform independent way.
}

