% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{extractTestInfo.BSGD}
\alias{extractTestInfo.BSGD}
\title{Extract test information (error rate) from command-line output}
\usage{
\method{extractTestInfo}{BSGD}(x, output)
}
\arguments{
\item{x}{object}

\item{output}{command-line output from BSGD}
}
\description{
Extract test information (error rate) from command-line output
}

