% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{extractTestInfo.LIBSVM}
\alias{extractTestInfo.LIBSVM}
\title{extractTestInfo.LIBSVM}
\usage{
\method{extractTestInfo}{LIBSVM}(x, output)
}
\arguments{
\item{x}{svm object}

\item{output}{}
}
\value{
error		error value
}
\description{
extractTestInfo.LIBSVM
}

