% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{extractTrainingInfo.BSGD}
\alias{extractTrainingInfo.BSGD}
\title{Extract training information (error rate) from command-line output}
\usage{
extractTrainingInfo.BSGD(x, output)
}
\arguments{
\item{x}{object}

\item{output}{command-line output from BSGD}
}
\description{
Extract training information (error rate) from command-line output
}

