% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{extractTrainingInfo.LIBSVM}
\alias{extractTrainingInfo.LIBSVM}
\title{extractTrainingInfo.LIBSVM}
\usage{
extractTrainingInfo.LIBSVM(x, output)
}
\arguments{
\item{x}{svm object}

\item{output}{}
}
\value{
error		error value
}
\description{
extractTrainingInfo.LIBSVM
}

