% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllSVMSoftware.R
\name{findAllSVMSoftware}
\alias{findAllSVMSoftware}
\title{findAllSVMSoftware 
	Given a search path, it will try to find the software packages
	for all registered SVM package by calling the 'findSoftware' routine of each known wrapper.}
\usage{
findAllSVMSoftware(searchPath = NA, verbose = FALSE)
}
\arguments{
\item{searchPath}{search the given path for the SVM binaries of all known SVM packages.}

\item{verbose}{print messages while searching?}
}
\description{
findAllSVMSoftware 
	Given a search path, it will try to find the software packages
	for all registered SVM package by calling the 'findSoftware' routine of each known wrapper.
}
\note{
To make sure that the binary is correct, it will be executed! (see findBinary for more infos)

If multiple binaries are found, the last one will be taken. Overwrite by hand, if necessary.
}

