% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSVMWrapper.R
\name{findAllSVMWrapper}
\alias{findAllSVMWrapper}
\title{findAllSVMWrapper
	given a search path, it will try to find the corresponding wrapper
	for all registered SVM packages.}
\usage{
findAllSVMWrapper(searchPath = NA, verbose = FALSE)
}
\arguments{
\item{searchPath}{search the given path for the SVM wrappers of all known SVM packages.}

\item{verbose}{print messages while searching?}
}
\description{
findAllSVMWrapper
	given a search path, it will try to find the corresponding wrapper
	for all registered SVM packages.
}
\note{
There is no extra check on the found wrapper. It will simply be sourced.

If multiple binaries are found, the last one will be taken. Overwrite by hand, if necessary.
}

