% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findBinary.R
\name{findBinary}
\alias{findBinary}
\title{findBinary
		given a file pattern and an std-out-pattern it will try to find a binary matching
		the file pattern (which contains std-out-pattern as output if called without arguments)
		in the path given.}
\usage{
findBinary(searchPath, pattern, outputPattern, verbose = FALSE)
}
\arguments{
\item{searchPath}{search the given path for the SVM binaries recursively.}

\item{pattern}{pattern for the binary file}

\item{outputPattern}{pattern for the stdout output of the binary}

\item{verbose}{print messages while searching?}
}
\value{
name of the found binary matching the pattern
}
\description{
findBinary
		given a file pattern and an std-out-pattern it will try to find a binary matching
		the file pattern (which contains std-out-pattern as output if called without arguments)
		in the path given.
}
\note{
To make sure that the binary is correct, it will be executed!

If multiple binaries are found, the last one will be taken. Overwrite by hand, if necessary.
}

