% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBinaryInDirectory.R
\name{findBinaryInDirectory}
\alias{findBinaryInDirectory}
\title{Find a given binary in given directory.}
\usage{
findBinaryInDirectory(binaryName = NULL, dir = NULL, patterns = NULL,
  verbose = FALSE)
}
\arguments{
\item{binaryName}{Search name for the SVM binary}

\item{dir}{Search path for the SVM binary.}

\item{patterns}{Pattern for output of the binary file}

\item{verbose}{Print messages while searching?}
}
\value{
name of the found binary matching the pattern
}
\description{
Given a file pattern and an std-out-pattern it will try to find a binary matching
the file pattern (which contains std-out-pattern as output if called without arguments)
in the path given.
}
\note{
To make sure that the binary is correct, it will be executed!
				This happens only if this routine is given some pattterns.

If multiple binaries are found, the last one will be taken. Overwrite by hand, if necessary.
}

