% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{findSoftware.BSGD}
\alias{findSoftware.BSGD}
\title{findSoftware.BSGD}
\usage{
\method{findSoftware}{BSGD}(x, searchPath = "./", verbose = FALSE)
}
\arguments{
\item{x}{SVM object}

\item{searchPath}{path to search for software}

\item{verbose}{print messages?}
}
\value{
x			binary path for the software
}
\description{
findSoftware.BSGD
}

