% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{findSoftware.LIBSVM}
\alias{findSoftware.LIBSVM}
\title{findSoftware.LIBSVM}
\usage{
\method{findSoftware}{LIBSVM}(x, searchPath = "./", verbose = FALSE)
}
\arguments{
\item{x}{svm object}

\item{searchPath}{path to search for software}

\item{verbose}{be verbose?}
}
\value{
x			svm object
}
\description{
findSoftware.LIBSVM
}

