% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSVMSoftware.R
\name{findSVMSoftware}
\alias{findSVMSoftware}
\title{Find software  for an SVM package.}
\usage{
findSVMSoftware(method = NA, searchPath = NA, verbose = FALSE)
}
\arguments{
\item{method}{name of the SVM method}

\item{searchPath}{search the given path for the SVM binaries of the given SVM method.}

\item{verbose}{print messages while searching?}
}
\description{
Given a search path, it will try to find the corresponding software packages
for the given method.
}
\note{
calls the findSoftware routine of the corresponding wrapper.
}

