% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateExamples.R
\name{generateConcentratedInformationExample}
\alias{generateConcentratedInformationExample}
\title{Generate data set with all information stored in the first two components}
\usage{
generateConcentratedInformationExample(filePath = NULL, n = 50,
  eps = 0.01, seed = NULL, normalize = TRUE, verbose = FALSE)
}
\arguments{
\item{filePath}{where to save the example file (as sparse data)}
}
\description{
Following the paper comparing nystrom to random fourier features
we generate 102 dimensional vectors, where the whole information is
in the first two components, the rest is uniformly sampled.
TODO: cite paper properly
}

