% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMWrapper.R
\name{getSupportedSVMProperties}
\alias{getSupportedSVMProperties}
\title{Return supported SVM properties.}
\usage{
getSupportedSVMProperties(type = NA_character_)
}
\arguments{
\item{type}{Type of SVM. Will be currently ignored!}
}
\description{
Return supported SVM properties.
}
\note{
type parameter is currently ignored!
}

