% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSVMPackage.R
\name{getSVMInstance}
\alias{getSVMInstance}
\title{dump specific package info}
\usage{
getSVMInstance(method = method)
}
\arguments{
\item{method}{name of package/method}
}
\value{
SVM object for the given method
}
\description{
dump specific package info
}

