% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSVMMethodsAsList.R
\name{getSVMMethodsAsList}
\alias{getSVMMethodsAsList}
\title{getSVMMethodsAsList}
\usage{
getSVMMethodsAsList(verbose = FALSE)
}
\arguments{
\item{verbose}{be verbose in output?}
}
\value{
A vector of all known SVM solvers as string.
}
\description{
Retrieve the names all known packages as a list of strings.
}

