% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSVMObject.R
\name{getSVMObject}
\alias{getSVMObject}
\title{getSVMObject}
\usage{
getSVMObject(method = method)
}
\arguments{
\item{method}{name of method}
}
\value{
SVM object for the given method (or NULL if it does not exist)
}
\description{
Retrieve the object for a given method.
}

