% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSVMParamSet.R
\name{getSVMParamSet}
\alias{getSVMParamSet}
\title{getSVMParamSet}
\usage{
getSVMParamSet(method)
}
\arguments{
\item{method}{Name of SVM method}
}
\value{
Parameters of thet SVM object for the given method (or NULL if it does not exist)
}
\description{
Retrieve the set of parameters for a given method.
}

