% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizationValuesLIBSVM.R
\name{optimizationValuesLIBSVM}
\alias{optimizationValuesLIBSVM}
\title{optimizationValues
	compute different optimization values for a given model}
\usage{
optimizationValuesLIBSVM(X, Y, model, C = 0, verbose = FALSE)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{model}{}

\item{C}{regularization parameter (if needed)}

\item{verbose}{report certain messages?}

\item{values}{which values to evaluate}
}
\value{
a list of values and corresponding objects
}
\description{
optimizationValues
	compute different optimization values for a given model
}

