% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSVMPackage.R
\name{outputAllSVMSoftwarePackages}
\alias{outputAllSVMSoftwarePackages}
\title{Dump all known package infos}
\usage{
outputAllSVMSoftwarePackages(verbose = TRUE)
}
\arguments{
\item{verbose}{be verbose in output?}
}
\value{
A vector of all known SVM solvers as string.
}
\description{
Dump all known package infos
}

