% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readLIBSVMModel.R
\name{readLIBSVMModel}
\alias{readLIBSVMModel}
\title{readLIBSVMModel
		Read LIBSVM model}
\usage{
readLIBSVMModel(modelFile = "./model", verbose = FALSE)
}
\arguments{
\item{modelFile}{model file to read}

\item{verbose}{be verbose?}
}
\value{
model object
}
\description{
readLIBSVMModel
		Read LIBSVM model
}
\note{
As this is a basic for all other model readers, we export it.
}

