% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{readModel.BSGD}
\alias{readModel.BSGD}
\title{readModel.BSGD}
\usage{
\method{readModel}{BSGD}(x, modelFile = "./model", verbose = FALSE,
  singleBinaryColumn = FALSE)
}
\arguments{
\item{x}{SVM object}

\item{modelFile}{path to model, defaults to a temporary file (given by R).}

\item{verbose}{print messages?}
}
\value{
svmatrix		BSGD model
}
\description{
readModel.BSGD
}

