% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{readModel.LIBSVM}
\alias{readModel.LIBSVM}
\title{Read LIBSVM model}
\usage{
\method{readModel}{LIBSVM}(x, modelFile = "./model", verbose = FALSE)
}
\arguments{
\item{x}{svm object}

\item{modelFile}{model file to read}

\item{verbose}{be verbose?}
}
\value{
model object
}
\description{
As this is a basic for all other model readers, we export it.
}

