% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readModel.R
\name{readModel}
\alias{readModel}
\title{readModel}
\usage{
readModel(modelType = NULL, modelFile = NULL, verbose = TRUE, ...)
}
\arguments{
\item{modelFile}{path to file to read model from}

\item{verbose}{be verbose}

\item{method}{name of the SVM method/solver}
}
\value{
SVM model
}
\description{
readModel
}

