% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readModelFromFile.R
\name{readModelFromFile}
\alias{readModelFromFile}
\title{readModel}
\usage{
readModelFromFile(modelFile = NULL, modelType = NULL, verbose = FALSE,
  ...)
}
\arguments{
\item{modelFile}{path to file to read model from}

\item{verbose}{be verbose}

\item{method}{name of the SVM method/solver}
}
\value{
SVM model
}
\description{
readModel
}

