% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{readPredictions.BSGD}
\alias{readPredictions.BSGD}
\title{Read predictions from BSGD}
\usage{
\method{readPredictions}{BSGD}(x, predictionsFilePath = "", verbose = FALSE)
}
\arguments{
\item{predictionsFile}{file to read predictions from}
}
\value{
array consisting of predictions
}
\description{
Read predictions from BSGD
}

