% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{readPredictions.LIBSVM}
\alias{readPredictions.LIBSVM}
\title{readPredictions.LIBSVM}
\usage{
\method{readPredictions}{LIBSVM}(x, predictionsFile = "", verbose = FALSE)
}
\arguments{
\item{x}{svm object}

\item{predictionsFile}{file to read predictions from}

\item{verbose}{be verbose?}
}
\value{
array consisting of predictions
}
\description{
readPredictions.LIBSVM
}

