% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseFormat.R
\name{readSparseFile}
\alias{readSparseFile}
\title{readSparseFile
	reads sparse data from a file}
\usage{
readSparseFile(fileName = "", verbose = FALSE)
}
\arguments{
\item{fileName}{path to the file}

\item{verbose}{report helpful messages?}
}
\description{
readSparseFile
	reads sparse data from a file
}

