% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSparseFileE1071.R, R/readSparseFormatE1071.R
\name{readSparseFileE1071}
\alias{readSparseFileE1071}
\title{Read Sparse File via E1071.}
\usage{
readSparseFileE1071(fileName = "", verbose = FALSE)

readSparseFileE1071(fileName = "", verbose = FALSE)
}
\arguments{
\item{fileName}{path to the file}

\item{verbose}{report helpful messages?}

\item{fileName}{path to the file}

\item{verbose}{report helpful messages?}
}
\description{
Reads sparse data from a file via e1071 library and return it as a dense matrix/vector.

readSparseFile
	reads sparse data from a file
}

