% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSparseFormatE1071.R, R/sparseFormat.R
\name{readSparseFormat}
\alias{readSparseFormat}
\title{readSparseFormat}
\usage{
readSparseFormat(con)

readSparseFormat(con)
}
\arguments{
\item{con}{connection to read from}

\item{con}{connection to read from}
}
\value{
list of supportvectors, coefficients and weights

list of supportvectors, coefficients and weights
}
\description{
readSparseFormat

readSparseFormat
}

