% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSparseFormatR.R
\name{readSparseFormatR}
\alias{readSparseFormatR}
\title{Read SparseFormat via R.}
\usage{
readSparseFormatR(con)
}
\arguments{
\item{con}{connection to read from}
}
\value{
list of supportvectors, coefficients and weights
}
\description{
Will read a sparse format file via R routines. Will nearly always be slower than the C++ method.
}

