% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSVMObject.R
\name{setSVMObject}
\alias{setSVMObject}
\title{setSVMObject}
\usage{
setSVMObject(method = method, object = NULL, verbose = FALSE)
}
\arguments{
\item{method}{name of method}

\item{object}{object to set}
}
\description{
Set a given object for a given method.
}

