% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SVMBridge.R
\docType{package}
\name{SVMBridge}
\alias{SVMBridge}
\alias{SVMBridge-package}
\title{SVMBridge}
\description{
An easy R interface to various SVM solvers.
}
\details{
This packages provides an interface to various SVM solvers like
LIBSVM, SVMperf, BudgetedSVM etc. These must be precompiled.
SVMBridge can find these automatically.
}

