% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SVMperfTrainingParameterCallBack}
\alias{SVMperfTrainingParameterCallBack}
\title{SVMperfTrainingParameterCallBack}
\usage{
SVMperfTrainingParameterCallBack(trainfile = "", modelFile = "",
  extraParameter = "", cost = 1, gamma = 1, kernelCacheSize = 1024,
  k = 1000, epsilon = 0.001, ...)
}
\arguments{
\item{...}{parameters that will be passed on training and/or test callbacks}
}
\description{
SVMperfTrainingParameterCallBack
}

