% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSVM.R
\name{testSVM}
\alias{testSVM}
\title{test a trained SVM.}
\usage{
testSVM(method = NULL, testDataFile = NULL, testDataX = NULL,
  testDataY = NULL, model = NULL, modelFile = NULL,
  readPredictions = FALSE, predictionsFile = NULL, extraParameter = "",
  verbose = FALSE, ...)
}
\arguments{
\item{testDataFile}{file to read training data from}

\item{testDataX}{matrix to read training data from}

\item{testDataY}{matrix to read training label from}

\item{model}{list containing a trained SVM model}

\item{modelFile}{path to a model file}

\item{readPredictions}{if TRUE, the prediction will be read back}

\item{predictionsFile}{the prediction will be written to this file or to a tempfile}

\item{extraParameter}{extra parameters for solver}

\item{verbose}{be verbose?}
}
\value{
SVM Object
}
\description{
This is the main routine for testing/prediction an SVM.
}
\note{
testDataFile and testDataX,testDataY are exclusive parameters, i.e. you cannot specify both.

Make sure the modelFile is a character string and not a factor! Else you might obtain strange modelfile arguments.
}

