% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trainSVM.R
\name{trainSVM}
\alias{trainSVM}
\title{trainSVM}
\usage{
trainSVM(method = NULL, trainDataFile = NULL, trainDataX = NULL,
  trainDatay = NULL, trainBinaryPath = NULL, extraParameter = "",
  modelFile = NULL, verbose = FALSE, timeOut = -1,
  readModelFile = FALSE, ...)
}
\arguments{
\item{method}{name of the SVM method/solver}

\item{trainDataFile}{file to read training data from}

\item{trainDataX}{matrix to read training data from}

\item{trainDatay}{matrix to read training label from}

\item{trainBinaryPath}{full path to the training binary to call}

\item{extraParameter}{extra parameters for solver}

\item{modelFile}{path to a model file}

\item{verbose}{be verbose?}

\item{model}{list containing a trained SVM model}

\item{trainingParameterCallBack}{function that returns command line for training}

\item{extractInformationCallBack}{function that extracts information from testing output}

\item{readModelCallBack}{function that reads a model into a list}
}
\value{
SVM Object
}
\description{
trainSVM
}
\note{
exclusive parameters, i.e. you cannot specify both:
}

