% Generated by roxygen2 (4.0.2): do not edit by hand
\name{universalSVMTrainer}
\alias{universalSVMTrainer}
\title{trainSVM}
\usage{
universalSVMTrainer(trainDataFile = NULL, trainDataX = NULL,
  trainDatay = NULL, trainBinaryPath = NULL, method = NULL,
  extraParameter = "", modelFilePath = NULL, verbose = FALSE,
  trainingParameterCallBack = trainingParameterCallBack(),
  extractInformationCallBack = extractInformationCallBack(),
  readModelCallBack = readModelCallBack(), ...)
}
\arguments{
\item{trainDataFile}{file to read training data from}

\item{trainDataX}{matrix to read training data from}

\item{trainDatay}{matrix to read training label from}

\item{trainBinaryPath}{full path to the training binary to call}

\item{method}{name of the SVM method/solver}

\item{extraParameter}{extra parameters for solver}

\item{verbose}{be verbose?}

\item{trainingParameterCallBack}{function that returns command line for training}

\item{extractInformationCallBack}{function that extracts information from testing output}

\item{readModelCallBack}{function that reads a model into a list}

\item{model}{list containing a trained SVM model}

\item{modelFile}{path to a model file}
}
\value{
SVM Object
}
\description{
trainSVM
}
\note{
exclusive parameters, i.e. you cannot specify both:
}
\examples{
universalSVMTrainer(trainDataFile = './data/australian.sparse')
}

