% Generated by roxygen2 (4.0.2): do not edit by hand
\name{universalWrapper}
\alias{universalWrapper}
\title{universalWrapper}
\usage{
universalWrapper(trainDataFile = NULL, testDataFile = NULL,
  trainDataX = NULL, trainDatay = NULL, testDataX = NULL,
  testDatay = NULL, trainBinaryPath = NULL, testBinaryPath = NULL,
  method = NULL, extraParameter = "", modelFilePath = NULL,
  model = NULL, verbose = FALSE,
  trainingParameterCallBack = trainingParameterCallBack(),
  testParameterCallBack = testParameterCallBack(),
  extractInformationCallBack = extractInformationCallBack(),
  readModelCallBack = readModelCallBack(),
  writeModelCallBack = writeModelCallBack(),
  predictionsCallBack = predictionsCallBack(), ...)
}
\arguments{
\item{trainDataFile}{file to read training data from}

\item{testDataFile}{file to read test data from}

\item{trainDataX}{matrix to read training data from}

\item{trainDatay}{matrix to read training label from}

\item{testDataX}{matrix to read test data from}

\item{testDatay}{matrix to read test label from}

\item{trainBinaryPath}{full path to the training binary to call}

\item{testBinaryPath}{full path to the test binary to call}

\item{method}{name of the SVM method/solver}

\item{extraParameter}{extra parameters for solver}

\item{model}{list containing a trained SVM model}

\item{verbose}{be verbose?}

\item{trainingParameterCallBack}{function that returns command line for training}

\item{testParameterCallBack}{function that returns command line for testing}

\item{extractInformationCallBack}{function that extracts information from testing output}

\item{readModelCallBack}{function that reads a model into a list}

\item{writeModelCallBack}{function that writes a model to a file}

\item{predictionsCallBack}{function that reads predictions from a file to a matrix}

\item{modelFile}{path to a model file}
}
\description{
universalWrapper
}
\note{
exclusive parameters, i.e. you cannot specify both:
}

