% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLIBSVMModel.R
\name{writeLIBSVMModel}
\alias{writeLIBSVMModel}
\title{Write LIBSVM model.}
\usage{
writeLIBSVMModel(model = NA, modelFile = "./model", verbose = FALSE)
}
\arguments{
\item{model}{model object to write}

\item{modelFile}{path where to write the model}

\item{verbose}{be verbose?}
}
\description{
Given a model in memory, write it out in LIBSVM model format, so that e.g. LIBSVM can predict on it directly.
}
\note{
As this is a basic for all other model readers, we export it.
}

