% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{writeModel.BSGD}
\alias{writeModel.BSGD}
\title{writeModel.BSGD}
\usage{
\method{writeModel}{BSGD}(x, model = NA, modelFile = "./model",
  verbose = FALSE)
}
\arguments{
\item{x}{SVM object}

\item{model}{Model to write}

\item{modelFile}{Path to write model to}

\item{verbose}{print messages?}
}
\description{
writeModel.BSGD
}

