% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIBSVM_wrapper.R
\name{writeModel.LIBSVM}
\alias{writeModel.LIBSVM}
\title{Write LIBSVM model}
\usage{
\method{writeModel}{LIBSVM}(x, model = NA, modelFile = "./model",
  verbose = FALSE)
}
\arguments{
\item{x}{svm object}

\item{model}{model object to write}

\item{modelFile}{path where to write the model}

\item{verbose}{be verbose?}
}
\description{
As this is a basic for all other model readers, we export it.
}

