% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModel.R
\name{writeModelToFile}
\alias{writeModelToFile}
\title{writeModelToFile}
\usage{
writeModelToFile(model = NULL, modelFile = NULL, verbose = FALSE, ...)
}
\arguments{
\item{modelFile}{path to file to write model to}

\item{verbose}{be verbose}
}
\value{
nothing
}
\description{
writeModelToFile
}

