% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeSparseData}
\alias{writeSparseData}
\title{Write given (dense) R matrix and R vector in sparse (LIBSVM) format to given file.}
\usage{
writeSparseData(filename, X, Y, append = FALSE, skipBytes = 0L,
  verbose = FALSE, zeroBased = FALSE)
}
\arguments{
\item{filename}{the filename to write the given data to}

\item{X}{The data to write as a matrix}

\item{Y}{The labels for the data.}

\item{append}{Append to the file (or overwrite)?}

\item{skipBytes}{How many bytes of the head of the files should be skipped?}

\item{verbose}{Be verbose?}

\item{zeroBased}{Do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}
}
\value{
NULL.
}
\description{
Write given (dense) R matrix and R vector in sparse (LIBSVM) format to given file.
}
\note{
Labels can any numeric, they are not converted in any way.
}
\examples{
	X = as.matrix(iris[,1:4])
	Y = as.matrix(as.numeric(iris[,5]))
	writeSparseData ("./australian.data", X, Y)

}
\keyword{IO}

