% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseFormat.R
\name{writeSparseDataToConnection}
\alias{writeSparseDataToConnection}
\title{writeSparseDataToConnection}
\usage{
writeSparseDataToConnection(con, X, Y, verbose = FALSE, zeroBased = FALSE)
}
\arguments{
\item{con}{connection}

\item{X}{input data}

\item{Y}{labels}

\item{verbose}{be verbose?}

\item{zeroBased}{do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}
}
\description{
writeSparseDataToConnection
}

