% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkModel.R
\name{checkModel}
\alias{checkModel}
\title{Check model.}
\usage{
checkModel(model = NA, verbose = FALSE)
}
\arguments{
\item{model}{model object to write}

\item{verbose}{be verbose?}

\item{modelFile}{path where to write the model}
}
\description{
Given a model in memory, write it out in LIBSVM model format, so that e.g. LIBSVM can predict on it directly.
}
\note{
As this is a basic for all other model readers, we export it.
}

