% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeOptimizationValues}
\alias{computeOptimizationValues}
\title{Compute several optimization values from a model like primal value}
\usage{
computeOptimizationValues(X, Y, C, gamma, SV, nSV, alpha, rho, label,
  verbose = FALSE)
}
\arguments{
\item{X}{input data}

\item{Y}{labels}

\item{C}{regularization value from svm problem}

\item{gamma}{used kernel bandwidth}

\item{SV}{matrix of support vectors}

\item{nSV}{vector of number of support vectors, needed for the multiclass case}

\item{rho}{vector of offset/bias terms}

\item{label}{vector of label mapping}

\item{alpa}{alpha of support vectors}
}
\value{
list of optimization values
}
\description{
Compute several optimization values from a model like primal value
}
\note{
works with LIBSVMlike models only

all computations will be done in DENSE format, not sparse!
}

