% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SVMWrapper.R
\name{createSVMWrapper}
\alias{createSVMWrapper}
\title{Description...}
\usage{
createSVMWrapper(method, trainingBinary = "", ...)
}
\arguments{
\item{a}{Description}

\item{b}{Description
...}
}
\description{
Details
}

