% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectModelTypeFromFile.R
\name{detectModelTypeFromFile}
\alias{detectModelTypeFromFile}
\title{detectModelTypeFromFile
		auto detect type of SVM model}
\usage{
detectModelTypeFromFile(modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{modelFile}{path to model file}

\item{verbose}{be verbose?}
}
\value{
method name or NULL if no method could be detected.
}
\description{
detectModelTypeFromFile
		auto detect type of SVM model
}

