% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectModelTypeFromFile.R
\name{detectModelTypeFromFile}
\alias{detectModelTypeFromFile}
\title{Detect model type from file.}
\usage{
detectModelTypeFromFile(modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{modelFile}{path to model file}

\item{verbose}{be verbose?}
}
\value{
method name or NULL if no method could be detected.
}
\description{
This will auto detect the type of SVM model from a given file.
}

