% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectMethodFromFile.R
\name{detectModelTypeFromFile}
\alias{detectModelTypeFromFile}
\title{Auto detect type of SVM model}
\usage{
detectModelTypeFromFile(filePath = NULL, verbose = FALSE)
}
\arguments{
\item{..}{}
}
\value{
method name or NULL if no method could be detected.
}
\description{
Auto detect type of SVM model
}

