% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectModelTypeFromFile.R
\name{detectModelTypeFromFile}
\alias{detectModelTypeFromFile}
<<<<<<< HEAD
\title{detectModelTypeFromFile
		auto detect type of SVM model}
=======
\title{Detect model type from file.}
>>>>>>> e37ade4654a469965a1b7d53c236dcf03510b207
\usage{
detectModelTypeFromFile(modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{modelFile}{path to model file}

\item{verbose}{be verbose?}
}
\value{
method name or NULL if no method could be detected.
}
\description{
<<<<<<< HEAD
detectModelTypeFromFile
		auto detect type of SVM model
=======
This will auto detect the type of SVM model from a given file.
>>>>>>> e37ade4654a469965a1b7d53c236dcf03510b207
}

