% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpSparseFormat.R
\name{dumpSparseFormat}
\alias{dumpSparseFormat}
\title{Dump SparseFormat.}
\usage{
dumpSparseFormat(labels, data)
}
\arguments{
\item{labels}{container for labels}

\item{data}{container for data}
}
\value{
sparseLine		string with all collected data
}
\description{
Simple (debug) dump of a sparse format given by labels and data into a string.
}

