% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpSparseFormat.R, R/sparseFormat.R
\name{dumpSparseLine}
\alias{dumpSparseLine}
\title{dumpSparseLine}
\usage{
dumpSparseLine(row, dense = FALSE)

dumpSparseLine(row, dense = FALSE)
}
\arguments{
\item{row}{array to read data from}

\item{row}{array to read data from}
}
\value{
sparseLine		string with row data

sparseLine		string with row data
}
\description{
dumpSparseLine

dumpSparseLine
}

