% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSVMSoftware.R
\name{findAllSVMSoftware}
\alias{findAllSVMSoftware}
\title{findAllSVMSoftware
		given a search path, it will try to find the corresponding software packages
		for all registered SVM packages.}
\usage{
findAllSVMSoftware(searchPath = NA, verbose = FALSE)
}
\arguments{
\item{searchPath}{search the given path for the SVM binaries of all known SVM packages.}

\item{verbose}{print messages while searching?}
}
\description{
findAllSVMSoftware
		given a search path, it will try to find the corresponding software packages
		for all registered SVM packages.
}
\note{
To make sure that the binary is correct, it will be executed! (see findBinary for more infos)

If multiple binaries are found, the last one will be taken. Overwrite by hand, if necessary.
}

