% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBinaryInDirectory.R
\name{findBinaryInDirectory}
\alias{findBinaryInDirectory}
\title{findBinaryInDirectory.
	given a file pattern and an std-out-pattern it will try to find a binary matching
	the file pattern (which contains std-out-pattern as output if called without arguments)
	in the path given.}
\usage{
findBinaryInDirectory(binaryName = NULL, dir = NULL, patterns = NULL,
  applyKeyFix = FALSE, verbose = FALSE)
}
\arguments{
\item{binaryName}{search the given path for the SVM binaries recursively.}

\item{dir}{search the given path for the SVM binaries recursively.}

\item{patterns}{pattern for the binary file}

\item{applyKeyFix}{apply fix for 'press any key to continue'?}

\item{verbose}{print messages while searching?}
}
\value{
name of the found binary matching the pattern
}
\description{
findBinaryInDirectory.
	given a file pattern and an std-out-pattern it will try to find a binary matching
	the file pattern (which contains std-out-pattern as output if called without arguments)
	in the path given.
}
\note{
To make sure that the binary is correct, it will be executed!
				This happens only if this routine is given some pattterns.

If multiple binaries are found, the last one will be taken. Overwrite by hand, if necessary.
}

