% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSVMWrapper.R
\name{findSVMWrapper}
\alias{findSVMWrapper}
\title{Find a given SVM Wrapper inside a search path.}
\usage{
findSVMWrapper(method = NA, searchPath = NA, recursive = TRUE,
  source = TRUE, verbose = FALSE)
}
\arguments{
\item{method}{Name of the SVM method}

\item{searchPath}{Search the given path for the SVM binaries of all known SVM packages.}

\item{recursive}{Recursive search?}

\item{source}{Source the found wrapper?}

\item{verbose}{Print messages while searching?}
}
\description{
Given a search path, it will try to find the corresponding wrapper for the given method.
This routine will also source the file, if specified.
}

