% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSVMWrapper.R
\name{findSVMWrapper}
\alias{findSVMWrapper}
\title{findSVMWrapper
		given a search path, it will try to find the corresponding wrapper for the given method.}
\usage{
findSVMWrapper(method = NA, name = NA, searchPath = NA, verbose = FALSE)
}
\arguments{
\item{method}{name of the SVM method}

\item{name}{if name is given, the search pattern will not have _wrapper.R postfix}

\item{searchPath}{search the given path for the SVM binaries of all known SVM packages.}

\item{verbose}{print messages while searching?}
}
\description{
findSVMWrapper
		given a search path, it will try to find the corresponding wrapper for the given method.
}
\note{
To make sure that the binary is correct, it will be executed! (see findBinary for more infos)

If multiple binaries are found, the last one will be taken. Overwrite by hand, if necessary.
}

