% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimizationValues.R
\name{optimizationValues}
\alias{optimizationValues}
\title{compute different optimization values for a given model}
\usage{
optimizationValues(X, Y, model, C = 0, values = c(), verbose = FALSE)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{model}{}

\item{C}{regularization parameter (if needed)}

\item{values}{which values to evaluate}
}
\value{
a list of values and corresponding objects
}
\description{
compute different optimization values for a given model
}

