% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLIBSVMModel.R
\name{readLIBSVMModel}
\alias{readLIBSVMModel}
\title{Read a model in LIBSVM format}
\usage{
readLIBSVMModel(modelFile = "./model", verbose = FALSE)
}
\arguments{
\item{modelFile}{model file to read}

\item{verbose}{be verbose?}
}
\value{
model object
}
\description{
This will simply read a model in LIBSVM format. As this is the base of many
other solvers, this is part of the library (instead of putting it into a wrapper)
and it uses Rcpp to read the support vectors for speeding up.
}

