% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readLIBSVMPredictions.R
\name{readLIBSVMPredictions}
\alias{readLIBSVMPredictions}
\title{Read predictions from a file}
\usage{
readLIBSVMPredictions(x, predictionsFile = "", verbose = FALSE)
}
\arguments{
\item{predictionsFile}{file to read predictions from}

\item{verbose}{be verbose?}
}
\value{
array consisting of predictions
}
\description{
Read predictions produced by LIBSVM, i.e. each line one label.
}

