% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLIBSVMPredictions.R
\name{readLIBSVMPredictions}
\alias{readLIBSVMPredictions}
\title{readLIBSVMPredictions
		Read predictions produced by LIBSVM, i.e. each line one label.}
\usage{
readLIBSVMPredictions(x, predictionsFile = "", verbose = FALSE)
}
\arguments{
\item{x}{SVMObject}

\item{predictionsFile}{file to read predictions from}

\item{verbose}{be verbose?}
}
\value{
array consisting of predictions
}
\description{
readLIBSVMPredictions
		Read predictions produced by LIBSVM, i.e. each line one label.
}

