% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSGD_wrapper.R
\name{readModel.BSGD}
\alias{readModel.BSGD}
\title{readModel.BSGD}
\usage{
\method{readModel}{BSGD}(x, modelFile = "./model", verbose = FALSE)
}
\arguments{
\item{x}{SVM object}

\item{modelFile}{path to model, defaults to a temporary file (given by R).}

\item{verbose}{print messages?}
}
\value{
<<<<<<< HEAD
svmatrix		svmmatrix object
=======
svmatrix		BSGD model
>>>>>>> e37ade4654a469965a1b7d53c236dcf03510b207
}
\description{
readModel.BSGD
}

