% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readModelFromFile.R
\name{readModelFromFile}
\alias{readModelFromFile}
\title{Read a model from a given file}
\usage{
readModelFromFile(modelFile = NULL, modelType = NULL, verbose = FALSE,
  ...)
}
\arguments{
\item{modelFile}{Path to the file to read the model from}

\item{modelType}{Name of the SVM method/solver. If Null, detectModel will be called.}

\item{verbose}{Be verbose?}
}
\value{
An SVM model
}
\description{
Read a model from a given file
}
\note{
If no modelType is given, the model will be inferred from detectModelTypeFromFile.
If this fails, an error will be issued.
}

