% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readSparseData}
\alias{readSparseData}
\title{Read a given file in sparse (LIBSVM) format to dense R matrix and R vector.}
\usage{
readSparseData(filename, skipBytes = 0L, verbose = FALSE,
  zeroBased = FALSE)
}
\arguments{
\item{filename}{the filename of the data in sparse file format. The Tilde Character isn't supported for f filename!}

\item{verbose}{show verbose output?}

\item{zeroBased}{do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}
}
\value{
the data is read into an R matrix and an R vector, containing the data
				and the labels. note, that these are not in sparse format, but are dense.
}
\description{
Read a given file in sparse (LIBSVM) format to dense R matrix and R vector.
}
\note{
this routine is nearly a 1:1 adoptation from the LIBSVM original code.
}
\examples{
\dontrun{
	S = readSparseData("../../../SVMBridge/tests/data/australian.train")
	print (paste("Data has", nrow(S$X), "points."))
	print (paste("Labels are", unique(S$Y), "."))
}
}
\keyword{IO}

