% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparseFormat.R
\name{readSparseDataFromConnection}
\alias{readSparseDataFromConnection}
\title{read spase data from an open connection}
\usage{
readSparseDataFromConnection(con, verbose = FALSE, zeroBased = FALSE)
}
\arguments{
\item{con}{connection}

\item{verbose}{be verbose?}

\item{zeroBased}{features start with 0?}
}
\description{
read spase data from an open connection
}

