% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSparseDataFromConnection.R, R/readSparseFormatE1071.R, R/sparseFormat.R
\name{readSparseDataFromConnection}
\alias{readSparseDataFromConnection}
\title{Read spase data from an open connection.}
\usage{
readSparseDataFromConnection(con, verbose = FALSE, zeroBased = FALSE)

readSparseDataFromConnection(con, verbose = FALSE, zeroBased = FALSE)

readSparseDataFromConnection(con, verbose = FALSE, zeroBased = FALSE)
}
\arguments{
\item{con}{connection}

\item{verbose}{be verbose?}

\item{zeroBased}{do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}

\item{con}{connection}

\item{verbose}{be verbose?}

\item{zeroBased}{do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}

\item{con}{connection}

\item{verbose}{be verbose?}

\item{zeroBased}{do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}
}
\value{
L		list containing input data and labels

L		list containing input data and labels

L		list containing input data and labels
}
\description{
Read spase data from an open connection.

read spase data from an open connection

read spase data from an open connection
}

