% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparseFormat.R
\name{readSparseFormat}
\alias{readSparseFormat}
\title{readSparseFormat}
\usage{
readSparseFormat(con)
}
\arguments{
\item{con}{connection to read from}
}
\value{
list of supportvectors, coefficients and weights
}
\description{
readSparseFormat
}

