% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSVMObject.R
\name{setSVMObject}
\alias{setSVMObject}
\title{Set an SVM object into the environment.}
\usage{
setSVMObject(method = method, object = NULL, verbose = FALSE)
}
\arguments{
\item{method}{Name of method}

\item{object}{Object to set.}

\item{verbose}{Be verbose?}
}
\description{
Set an SVM object into the environment.
}

