% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testSVM.R
\name{testSVM}
\alias{testSVM}
\title{testSVM}
\usage{
testSVM(method = NULL, testDataFile = NULL, testDataX = NULL,
  testDatay = NULL, model = NULL, modelFile = NULL,
  readPredictions = FALSE, predictionsFile = NULL, extraParameter = "",
  verbose = FALSE, ...)
}
\arguments{
\item{method}{name of the SVM method/solver}

\item{model}{list containing a trained SVM model}

\item{modelFile}{path to a model file}

\item{extraParameter}{extra parameters for solver}

\item{verbose}{be verbose?}

\item{trainDataFile}{file to read training data from}

\item{trainDataX}{matrix to read training data from}

\item{trainDatay}{matrix to read training label from}

\item{trainBinaryPath}{full path to the training binary to call}

\item{trainingParameterCallBack}{function that returns command line for training}

\item{extractInformationCallBack}{function that extracts information from testing output}

\item{readModelCallBack}{function that reads a model into a list}
}
\value{
SVM Object
}
\description{
testSVM
}
\note{
exclusive parameters, i.e. you cannot specify both:
}

