% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainSVM.R
\name{trainSVM}
\alias{trainSVM}
\title{Train an SVM.}
\usage{
trainSVM(method = NULL, trainDataFile = NULL, trainDataX = NULL,
  trainDataY = NULL, trainBinaryPath = NULL, extraParameter = "",
  modelFile = NULL, subsamplingRate = NULL, subsamplingMethod = "cutoff",
  verbose = FALSE, timeOut = -1, readModelFile = FALSE, ...)
}
\arguments{
\item{method}{name of the SVM method/solver}

\item{trainDataFile}{Filename to read training data from. This cannot be used at the same time as trainDataX/trainDataY.}

\item{trainDataX}{Matrix comprising the data. This cannot be used at the same time as trainDataFile.}

\item{trainDataY}{Vector comprising the labels for the data. This cannot be used at the same time as trainDataFile.}

\item{trainBinaryPath}{Full path to the training binary to call.}

\item{extraParameter}{Extra parameters for solver, passed to the createTrainingArguments function of the wrapper.}

\item{modelFile}{Path to the model file to create. If none specified, a temporary file will be created and read into memory.}

\item{subsamplingRate}{Rate to subsample the data.}

\item{subsamplingMethod}{Method to subsample the data with.}

\item{verbose}{Be verbose?}

\item{timeOut}{Value to time out at, so that e.g. solvers that do stall/do an endless loop can be stopped nonetheless.}

\item{readModelFile}{if TRUE, will read the model back (model is saved in the created object)}

\item{...}{Further parameters that will be passed to the wrapper.}
}
\value{
An SVM Object.
}
\description{
This is the main routine that trains an SVM, e.g. it will make sure that all
data is on the disk, create the training arguments and then call the corresponding binary.
}
\note{
trainDataFile, trainDataX/trainDataY are exclusive parameters, i.e. you cannot specify both.

timeOut might not work on all platforms yet.
}

