% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainSVM.R
\name{trainSVM}
\alias{trainSVM}
\title{trainSVM}
\usage{
trainSVM(method = NULL, trainDataFile = NULL, trainDataX = NULL,
  trainDatay = NULL, trainBinaryPath = NULL, extraParameter = "",
  modelFile = NULL, verbose = TRUE, subsamplingRate = NULL,
  subsamplingMethod = "cutoff", timeOut = -1, readModelFile = FALSE, ...)
}
\arguments{
\item{method}{name of the SVM method/solver}

\item{trainDataFile}{file to read training data from}

\item{trainDataX}{matrix to read training data from}

\item{trainDatay}{matrix to read training label from}

\item{trainBinaryPath}{full path to the training binary to call}

\item{extraParameter}{extra parameters for solver}

\item{modelFile}{path to a model file}

\item{verbose}{be verbose}

\item{subsamplingRate}{rate to subsample the data at}

\item{subsamplingMethod}{method to subsample the data with}

\item{timeOut}{value to time out at}

\item{readModelFile}{if TRUE, will read the model back (model is saved in the created object)}
}
\value{
SVM Object
}
\description{
trainSVM
}
\note{
exclusive parameters, i.e. you cannot specify both:
}

