% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelToFile.R
\name{writeModelToFile}
\alias{writeModelToFile}
\title{Write a given moel to a file.}
\usage{
writeModelToFile(model = NULL, modelFile = NULL, verbose = FALSE)
}
\arguments{
\item{model}{Model to write}

\item{modelFile}{Path to file to write model to}

\item{verbose}{Be verbose?}
}
\value{
nothing
}
\description{
Write a given moel to a file.
}

