% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeSparseData}
\alias{writeSparseData}
\title{Write given (dense) R matrix and R vector in sparse (LIBSVM) format to given file.}
\usage{
writeSparseData(filename, X, Y, verbose = FALSE, zeroBased = FALSE)
}
\arguments{
\item{filename}{the filename to write the given data to}

\item{verbose}{show verbose output?}

\item{zeroBased}{do the indices in the file start with 0, e.g. -1 0:2 1:4 ...?}
}
\value{
NULL.
}
\description{
Write given (dense) R matrix and R vector in sparse (LIBSVM) format to given file.
}
\note{
labels can any numeric, they are not converted in any way.
}
\examples{
X = as.matrix(iris[,1:4])
	Y = as.matrix(as.numeric(iris[,5]))
	writeSparseData (X, Y, "./australian.data")
}
\keyword{IO}

